requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
blinker==1.7.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.3.2
cron-descriptor==1.4.0
Django==4.2.6
django-celery-beat==2.5.0
django-filter==23.3
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
drf-yasg==1.21.7
eventlet==0.33.3
Flask==3.0.0
greenlet==3.0.1
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.3
MarkupSafe==2.1.3
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==7.3.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.9
Werkzeug==3.0.1
